imporove rounding in reverse address interpolation
When matchSide needed to round an address to even or odd it would just
decrement or increment it. If a reverse address interpolation produced
1777.4 and the range was even it would round to 1776. After these
changes it rounds to the nearer even number, 1778.

It should return a number strictly within the range as long as both ends
are even for even ranges and both ends are odd for odd ranges. This is
the same as the previous rounding.