fix reverse address lookup for features  with multiple geometries
This should fix reverse address interpolation for features with more
than a single geometry in their geometry collections.

The `ranges` variable passed to `matchSide` used always be the last set
of ranges for that feature. `ranges` is was set in the for loop and
`matchSide` called it with whatever value it was last set to.

This passes the correct corresponding `range` to `matchSide`.